requirements.txt
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
bottle==0.12.19
click==8.0.3
decorator==5.1.1
distlib==0.3.4
executing==0.8.2
filelock==3.6.0
greenlet==1.1.2
ipython==8.0.1
ipython-genutils==0.2.0
ipython-sql==0.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prettytable==0.7.2
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
virtualenv==20.13.1
wcwidth==0.2.5
